+++ CHANGELOG.md
@@ -12,11 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Update the Incentives component to handle CMS data (#474)
 - Applies new local tokens to `ProductShelf` component (#464)
 
 ### Deprecated
 
 ### Removed
+- Removes CSS imports of components that are not being used (#476)
 
 ### Fixed
 - Potential layout shift on Hero section fixed (#472)